requirements.txt
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0
fugashi>=1.1.0
unidic>=1.3.0
pykakasi>=2.2.0
